DF-10193 fix compilation
Former-commit-id: 1a199d4e5975940f7328c362024a678be5c94e27

+++ Sources/Internal/FileSystem/DynamicMemoryFile.cpp
@@ -158,7 +158,7 @@ bool DynamicMemoryFile::Truncate(uint64 size)
     if (!(fileAttributes & File::WRITE))
         return false;
 
-    data.resize(size);
+    data.resize(size_t(size));
     currentPtr = Min(currentPtr, size);
     isEof = (currentPtr == size);
